Extract S7ValueConverter
